Add files via upload
The utils.py file is a collection of utility functions and helper code used throughout the ReLife app. It includes general-purpose functions that are not specific to any single aspect of the application, such as formatting data, handling custom CSS, or other common tasks. This file improves code reusability and separation of concerns by abstracting common functionality used in multiple places.